Update requirements.txt
Recently, PyMongo 4.0 was released, which removes the Collection.remove() method. This method is called in wsgi.py, and breaks loading of data. I'm not a Mongo expert, and I haven't looked at what replaces the remove() method, but a simple fix is to pin PyMongo at <4.0. This forces installation of 3.12.3 (as I write this) and makes data loading work again.